Change the polygons and add the httpClient
- make the polygons in the dict to real Polygons
- outsource the generate_random into another module
- add the httpClient to main and api.py
