+++ guess-the-country/country-backend/country/coordinates/polygon.py
@@ -1,0 +1,11 @@
+import random
+
+
+def generate_random(polygon):
+    points = []
+    minx, miny, maxx, maxy = polygon.bounds
+    while len(points) < 1:
+        pnt = Point(random.uniform(minx, maxx), random.uniform(miny, maxy))
+        if polygon.contains(pnt):
+            points.append(pnt)
+    return list(points[0].coords)
